moving to dot plots
